requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Experiment tracking
wandb>=0.15.0

# Data handling
roboflow>=1.1.0
albumentations>=1.3.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0
